code cleanup complete made script and files more efficient
